pyproject.toml
[tool.poetry]
authors = ["m09 <142691+m09@users.noreply.github.com>"]
description = "Tool to handle invoices."
homepage = "https://github.com/m09/invoicez"
license = "Apache-2.0"
name = "invoicez"
readme = "README.md"
version = "4.0.0"

[tool.poetry.dependencies]
gitpython = "^3"
google-api-python-client = "^2"
google-auth-oauthlib = "^1"
jinja2 = "^3"
pandas = "^2"
pydantic = "^2"
python = ">= 3.10, < 3.13"
pyyaml = "^6"
rich = "^13"
typer = "^0.9"

[tool.poetry.group.dev.dependencies]
black = "^24"
flake8 = "^7"
flake8-bugbear = "^24"
flake8-docstrings = "^1"
isort = "^5"
mypy = "^1"
pdfminer-six = "^20231228"
pylint = "^3"
pytest = "^8"
pytest-cov = "^4"
types-pyyaml = "^6"

[tool.poetry.scripts]
invoicez = "invoicez.cli:main"

[tool.isort]
profile = "black"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
